getting ready, rename old vault support, update readme, makefile, etc.
