+++ docker-compose/compose.yaml
@@ -43,7 +43,7 @@ services:
       start_period: 30s
 
   exporter:
-    image: container-registry.oracle.com/database/observability-exporter:1.5.5
+    image: container-registry.oracle.com/database/observability-exporter:1.6.0
     container_name: exporter
     # command:
       # - '--log.level=debug'
